Adicionando configuracao do SqlServer

+++ APIMarketList.Domain/Entities/Member.cs
@@ -5,8 +5,8 @@ namespace APIMarketList.Domain.Entities
     public class Member : BaseEntity
     {
         public bool IsAdmin { get; set; }
-        public long GroupId { get; set; }
-        public long UserId { get; set; }
+        public int GroupId { get; set; }
+        public int UserId { get; set; }
 
         public virtual Group? Group { get; set; }
         public virtual User? User { get; set; }
